aicodes/requirements.txt
# Web 框架
Flask==3.0.0
flask-cors==4.0.0

# 机器学习
xgboost==2.0.3
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.26.2

# 数据可视化
plotly==5.18.0
openpyxl==3.1.2

# API 客户端
requests==2.31.0

# 环境变量管理
python-dotenv==1.0.0

# AI 模型和语音合成
langchain-openai==0.1.0
dashscope>=1.25.0

# 阿里云语音合成 SDK (可选 - 如果使用旧版)
# alibabacloud_cosyvoice20220616==1.0.0
# alibabacloud_tea_openapi==0.3.8

# 工具
joblib==1.3.2
